Inheritance has been implemented for old Car class and new Bus class both.

+++ src/utilities/Utilities.java
@@ -5,7 +5,7 @@ public class Utilities {
     public static String validationAndDefaultSet(String input, String defaultValue) {
         return input == null || input.isBlank() ? defaultValue : input;
     }
-    public static double validationAndDefaultSet(double input, int defaultValue) {
+    public static double validationAndDefaultSet(double input, double defaultValue) {
         return input > 0 ? input : defaultValue;
     }
     public static int validationAndDefaultSet(int input, int defaultValue) {
